# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Here QA can test and frontend or mobile developers can get information of API endpoints.",
        "title": "Monolithic project API Endpoints",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/v1",
    "paths": {
...
        "/public/pipeline": {
            "post": {
                "description": "Here pipeline can be created.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Public Pipeline"
                ],
                "summary": "Create pipeline",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Access key of a project",
                        "name": "AccessKey",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Secret key of a project",
                        "name": "SecretKey",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "post info",
                        "name": "post",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreatePipelineIntegrationReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.PipelineApiResponse"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/models.DefaultResponse"
                        }
                    }
                }
            }
        },
        "/storage": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Here storage can be updated.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Storage"
                ],
                "summary": "Update storage",
                "parameters": [
                    {
                        "description": "post info",
                        "name": "post",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.StorageUpdateReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.StorageApiResponse"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/models.DefaultResponse"
                        }
                    }
                }
            }
        },
        "/storage/list": {
            "get": {
                "description": "Here all storages can be got.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Storage"
                ],
                "summary": "Get storages list",
                "parameters": [
                    {
                        "type": "integer",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "order_by_created_at",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.StorageApiFindResponse"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/models.DefaultResponse"
                        }
                    }
                }
            }
        },
        "/storage/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Here storage can be deleted.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Storage"
                ],
                "summary": "Delete storage",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.DefaultResponse"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/models.DefaultResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.AudioTrack": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "input_url": {
                    "type": "string"
                },
                "lang_code": {
                    "type": "string"
                },
                "language": {
                    "type": "string"
                },
                "pipeline_id": {
                    "type": "string"
                },
                "size_kb": {
                    "type": "number"
                }
            }
        },
        "models.AudioTrackRequest": {
            "type": "object",
            "properties": {
                "input_url": {
                    "type": "string"
                },
                "language": {
                    "type": "string"
                },
                "language_code": {
                    "type": "string"
                }
            }
        },
...
        "models.CreatePipelineIntegrationReq": {
            "type": "object",
            "properties": {
                "audio_tracks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.AudioTrackRequest"
                    }
                },
                "bucket_name": {
                    "type": "string"
                },
                "input_url": {
                    "type": "string"
                },
                "language": {
                    "type": "string"
                },
                "language_code": {
                    "type": "string"
                },
                "output_key": {
                    "type": "string"
                },
                "project_id": {
                    "type": "string"
                },
                "resolutions": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "subtitle": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.SubtitleRequest"
                    }
                }
            }
        },
        "models.DefaultResponse": {
            "type": "object",
            "properties": {
                "error_code": {
                    "type": "integer"
                },
                "error_message": {
                    "type": "string"
                }
            }
        },
...
        "models.PipelineApiResponse": {
            "type": "object",
            "properties": {
                "body": {
                    "$ref": "#/definitions/models.PipelineResponse"
                },
                "error_code": {
                    "type": "integer"
                },
                "error_message": {
                    "type": "string"
                }
            }
        },
...
        "models.PipelineResponse": {
            "type": "object",
            "properties": {
                "audio_tracks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.AudioTrack"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "fail_description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "input_url": {
                    "type": "string"
                },
                "language": {
                    "type": "string"
                },
                "language_code": {
                    "type": "string"
                },
                "max_resolution": {
                    "type": "string"
                },
                "output_key": {
                    "type": "string"
                },
                "output_path": {
                    "type": "string"
                },
                "preparation_duration_seconds": {
                    "type": "number"
                },
                "project_id": {
                    "type": "string"
                },
                "resolutions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Resolution"
                    }
                },
                "resolutions_string": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "size_kb": {
                    "type": "number"
                },
                "stage": {
                    "type": "string"
                },
                "stage_status": {
                    "type": "string"
                },
                "subtitle": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Subtitle"
                    }
                },
                "transcode_duration_secs": {
                    "type": "number"
                },
                "updated_at": {
                    "type": "string"
                },
                "upload_duration_secs": {
                    "type": "number"
                },
                "video_duration": {
                    "type": "number"
                },
                "webhook_last_retry": {
                    "type": "string"
                },
                "webhook_retry_count": {
                    "type": "integer"
                },
                "webhook_status": {
                    "type": "boolean"
                }
            }
        },
...
        "models.Resolution": {
            "type": "object",
            "properties": {
                "bitrate": {
                    "type": "string"
                },
                "measure": {
                    "type": "string"
                },
                "resolution": {
                    "type": "string"
                }
            }
        },
        "models.StorageApiFindResponse": {
            "type": "object",
            "properties": {
                "body": {
                    "$ref": "#/definitions/models.StorageFindResponse"
                },
                "error_code": {
                    "type": "integer"
                },
                "error_message": {
                    "type": "string"
                }
            }
        },
        "models.StorageApiResponse": {
            "type": "object",
            "properties": {
                "body": {
                    "$ref": "#/definitions/models.StorageResponse"
                },
                "error_code": {
                    "type": "integer"
                },
                "error_message": {
                    "type": "string"
                }
            }
        },
        "models.StorageFindResponse": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "storages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.StorageResponse"
                    }
                }
            }
        },
        "models.StorageResponse": {
            "type": "object",
            "properties": {
                "access_key": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "domain_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "region": {
                    "type": "string"
                },
                "secret_key": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.StorageUpdateReq": {
            "type": "object",
            "properties": {
                "access_key": {
                    "type": "string"
                },
                "domain_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "region": {
                    "type": "string"
                },
                "secret_key": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "models.Subtitle": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "input_url": {
                    "type": "string"
                },
                "language": {
                    "type": "string"
                },
                "language_code": {
                    "type": "string"
                },
                "pipeline_id": {
                    "type": "string"
                },
                "size_kb": {
                    "type": "number"
                }
            }
        },
        "models.SubtitleRequest": {
            "type": "object",
            "properties": {
                "input_url": {
                    "type": "string"
                },
                "language": {
                    "type": "string"
                },
                "language_code": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}